pyproject.toml
[tool.poetry]
name = "openstreetmap-calendar"
version = "2.2.0"
description = ""
authors = ["Thomas Skowron <th@skowron.eu>"]
license = "APACHE 2.0"

[tool.poetry.dependencies]
python = "^3.7"
psycopg2-binary = "~=2.8"
requests = "~=2.22"
requests-oauthlib = "~=1.2"
django-leaflet = "~=0.27"
Django = "~=3.2"
Markdown = "~=3.2"
bleach = "~=3.3"
gunicorn = "~=20.0"
python-dotenv = "~=0.14"
sentry-sdk = "~=1.5"
django-prometheus = "~=2.0"
timezonefinder = "~=4.4"
Babel = "~=2.9"
django-background-tasks = "*"
# django-background-tasks-updated = "==1.2.7" # pinned to a fork which supports Django 4

[tool.poetry.dev-dependencies]
PyYaml = "~=5.3"
pylint = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
